Move poolsense coordinator to its own file
