+++ .coveragerc
@@ -978,6 +978,7 @@ omit =
     homeassistant/components/point/sensor.py
     homeassistant/components/poolsense/__init__.py
     homeassistant/components/poolsense/binary_sensor.py
+    homeassistant/components/poolsense/coordinator.py
     homeassistant/components/poolsense/sensor.py
     homeassistant/components/powerwall/__init__.py
     homeassistant/components/progettihwsw/__init__.py
